backend/requirements.txt
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
pylint-django==2.3.0
